feat : PR 싱크 후 추가 로직 추가

+++ data/src/main/java/com/stop/data/remote/source/nearplace/NearPlaceRemoteDataSourceImpl.kt
@@ -12,8 +12,8 @@ internal class NearPlaceRemoteDataSourceImpl @Inject constructor(
     override suspend fun getNearPlaces(
         version: Int,
         searchKeyword: String,
-        centerLon: Float,
-        centerLat: Float,
+        centerLon: Double,
+        centerLat: Double,
         appKey: String
     ): List<Place> {
         val result = nearPlaceApiService.getNearPlaces(
