+++ data/src/main/java/com/stop/data/repository/NearPlaceRepositoryImpl.kt
@@ -14,8 +14,8 @@ internal class NearPlaceRepositoryImpl @Inject constructor(
     override suspend fun getNearPlaces(
         version: Int,
         searchKeyword: String,
-        centerLon: Float,
-        centerLat: Float,
+        centerLon: Double,
+        centerLat: Double,
         appKey: String
     ): Flow<List<Place>> = flow<List<Place>> {
         emit(
