+++ domain/src/main/java/com/stop/domain/usecase/nearplace/GetNearPlacesUseCaseImpl.kt
@@ -12,8 +12,8 @@ internal class GetNearPlacesUseCaseImpl @Inject constructor(
     override suspend fun getNearPlaces(
         version: Int,
         searchKeyword: String,
-        centerLon: Float,
-        centerLat: Float,
+        centerLon: Double,
+        centerLat: Double,
         appKey: String
     ): Flow<List<Place>> =
         nearPlaceRepository.getNearPlaces(
